Fixes NewLines
When Passage Processor runs, sometimes you get newlines that don't print right. It's like the JAVA one would start printing, and then the C one would interupt it after it put the string on the buffer, but not the newline character because I was using System.out.println(), but now I just add a \n to all of those and changed it to print(). Maybe that'll fix it?